moved deployments to clusterView
